% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paprpkgr-package.R
\docType{package}
\name{paprpkgr-package}
\alias{paprpkgr}
\alias{paprpkgr-package}
\title{paprpkgr: Install personal framework for research compendium}
\description{
This package will install the framework for a research compendium
    into a package directory. Please see the README for usage details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wmoldham/paprpkgr}
  \item Report bugs at \url{https://github.com/wmoldham/paprpkgr/issues}
}

}
\author{
\strong{Maintainer}: William Oldham \email{wm.oldham@gmail.com} (\href{https://orcid.org/0000-0003-3029-4866}{ORCID})

}
\keyword{internal}
